requirements.txt
asgiref==3.8.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
Django==5.2.1
django-cors-headers==4.7.0
django-dynamic-preferences==1.17.0
django-filter==25.1
django-jazzmin==3.0.1
django-users-auth==2.4.16
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
josepy==2.1.0
mozilla-django-oidc==4.0.1
persisting-theory==1.0
pillow==11.3.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
